Final Tweaks
Added some database content and made radio buttons self-select in edit
views.

+++ db/migrate/20141123190747_create_exercises.rb
@@ -2,7 +2,7 @@ class CreateExercises < ActiveRecord::Migration
   def change
     create_table :exercises do |t|
       t.integer :exercise_category
-      t.integer :duration
+      t.float :duration
       t.string :intensity
       t.string :favsave
       t.integer :user_id
