+++ db/migrate/20141208184738_add_calorie_to_beers.rb
@@ -1,5 +1,5 @@
 class AddCalorieToBeers < ActiveRecord::Migration
   def change
-    add_column :beers, :calorie, :integer
+    add_column :beers, :calorie, :float
   end
 end
