+++ db/migrate/20141208184911_add_calorie_to_exercises.rb
@@ -1,5 +1,5 @@
 class AddCalorieToExercises < ActiveRecord::Migration
   def change
-    add_column :exercises, :calorie, :integer
+    add_column :exercises, :calorie, :float
   end
 end
